+++ src/main/java/org/jenkinsci/plugins/ghprb/GhprbRepository.java
@@ -306,37 +306,6 @@ public class GhprbRepository {
         GhprbTrigger.getDscp().save();
     }
 
-    boolean checkSignature(String body, String signature) {
-        String secret = helper.getTrigger().getSecret();
-        if (secret != null && ! secret.isEmpty()) {
-            if (signature != null && signature.startsWith("sha1=")) {
-                String expected = signature.substring(5);
-                String algorithm = "HmacSHA1";
-                try {
-                    SecretKeySpec keySpec = new SecretKeySpec(secret.getBytes(), algorithm);
-                    Mac mac = Mac.getInstance(algorithm);
-                    mac.init(keySpec);
-                    byte[] localSignatureBytes = mac.doFinal(body.getBytes("UTF-8"));
-                    String localSignature = Hex.encodeHexString(localSignatureBytes);
-                    if (! localSignature.equals(expected)) {
-                        logger.log(Level.SEVERE, "Local signature {0} does not match external signature {1}",
-                                new Object[] {localSignature, expected});
-                        return false;
-                    }
-                } catch (Exception e) {
-                    logger.log(Level.SEVERE, "Couldn't match both signatures");
-                    return false;
-                }
-            } else {
-                logger.log(Level.SEVERE, "Request doesn't contain a signature. Check that github has a secret that should be attached to the hook");
-                return false;
-            }
-        }
-
-        logger.log(Level.INFO, "Signatures checking OK");
-        return true;
-    }
-
     @VisibleForTesting
     void setHelper(Ghprb helper) {
         this.helper = helper;
